refactor(chat): extract shared message parser, fix error handling in chat streams
- Extract duplicate parseMessageContent function from PersonaChat and PersonaChatInline into shared module
- Add error handling for non-string stream updates (step: ERROR objects)
- Fix chatWithPersonaAction to use stream.done instead of stream.error for graceful error delivery
- Change streamable value type from string to any to support structured error payloads
